+++ android/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.mygdx.game">
+    package="com.mygdx.mathdash">
 
     <uses-feature android:glEsVersion="0x00020000" android:required="true" />
 
@@ -14,7 +14,7 @@
         android:label="@string/app_name"
         tools:ignore="UnusedAttribute">
         <activity
-            android:name="com.mygdx.game.AndroidLauncher"
+            android:name="com.mygdx.mathdash.AndroidLauncher"
             android:label="@string/app_name" 
             android:screenOrientation="sensorLandscape"
             android:configChanges="keyboard|keyboardHidden|navigation|orientation|screenSize|screenLayout"
