+++ build.gradle
@@ -23,7 +23,7 @@ allprojects {
 
     version = '1.0'
     ext {
-        appName = "Geometry Dash 2.0"
+        appName = "MathDash"
         gdxVersion = '1.11.0'
         roboVMVersion = '2.3.16'
         box2DLightsVersion = '1.5'
