--- core/src/com/mygdx/game/Sprites/TileObjects/Thorns.java
@@ -1,22 +1,0 @@
-package com.mygdx.game.Sprites.TileObjects;
-
-import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.maps.tiled.TiledMap;
-import com.badlogic.gdx.math.Rectangle;
-import com.badlogic.gdx.physics.box2d.World;
-import com.mygdx.game.MathDash;
-
-public class Thorns extends InteractiveTileObject {
-
-    public Thorns(World world, TiledMap map, Rectangle bounds) {
-        super(world, map, bounds);
-        fixture.setUserData(this);
-        setCateforyFilter(MathDash.THORNS_BIT);
-
-    }
-
-    @Override
-    public void contacting() {
-        Gdx.app.log("Шипы", "Соприкосновение");
-    }
-}
