--- core/src/com/mygdx/game/db/Preference.java
@@ -1,16 +1,0 @@
-package com.mygdx.game.db;
-
-import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.Preferences;
-
-public class Preference {
-
-    public Preferences Attempt_Сounter;
-
-    public Preference() {
-
-        Attempt_Сounter = Gdx.app.getPreferences("game_Score");
-
-    }
-
-}
