+++ core/src/com/mygdx/mathdash/screens/GameFinishedScreen.java
@@ -1,0 +1,100 @@
+package com.mygdx.mathdash.screens;
+
+import com.badlogic.gdx.Game;
+import com.badlogic.gdx.Gdx;
+import com.badlogic.gdx.Screen;
+import com.badlogic.gdx.graphics.Color;
+import com.badlogic.gdx.graphics.GL20;
+import com.badlogic.gdx.graphics.OrthographicCamera;
+import com.badlogic.gdx.graphics.g2d.BitmapFont;
+import com.badlogic.gdx.scenes.scene2d.InputEvent;
+import com.badlogic.gdx.scenes.scene2d.Stage;
+import com.badlogic.gdx.scenes.scene2d.ui.Label;
+import com.badlogic.gdx.scenes.scene2d.ui.Table;
+import com.badlogic.gdx.scenes.scene2d.utils.ClickListener;
+import com.badlogic.gdx.utils.viewport.FitViewport;
+import com.badlogic.gdx.utils.viewport.Viewport;
+import com.mygdx.mathdash.MathDash;
+
+public class GameFinishedScreen implements Screen {
+    private Viewport viewport;
+    private Stage stage;
+
+    private Game game;
+
+    private Label menuBack;
+
+
+    public GameFinishedScreen(final MathDash game) {
+        this.game = game;
+        viewport = new FitViewport(MathDash.V_WIDTH, MathDash.V_HEIGHT, new OrthographicCamera());
+        stage = new Stage(viewport, (game).batch);
+
+
+        Gdx.input.setInputProcessor(stage);
+
+        Label.LabelStyle font = new Label.LabelStyle(new BitmapFont(), Color.WHITE);
+
+        Table table = new Table();
+        table.center();
+        table.setFillParent(true);
+
+        Label gameEndLabel = new Label("Level passed", font);
+        menuBack = new Label("Back to Menu", font);
+
+        table.add(gameEndLabel).expandX();
+        table.row();
+        table.add(menuBack).padTop(25f);
+
+        stage.addActor(table);
+
+        menuBack.addListener(new ClickListener() {
+            @Override
+            public void clicked(InputEvent event, float x, float y) {
+
+                game.setScreen(new MainMenu(game));
+                dispose();
+
+            }
+        });
+    }
+
+
+    @Override
+    public void show() {
+
+    }
+
+    @Override
+    public void render(float delta) {
+
+        Gdx.gl.glClearColor(.1f, .12f, .16f, 1);
+        Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT);
+        stage.draw();
+    }
+
+    @Override
+    public void resize(int width, int height) {
+
+    }
+
+    @Override
+    public void pause() {
+
+    }
+
+    @Override
+    public void resume() {
+
+    }
+
+    @Override
+    public void hide() {
+
+    }
+
+    @Override
+    public void dispose() {
+        stage.dispose();
+    }
+}
