+++ core/src/com/mygdx/mathdash/sprites/tileObjects/FinishBlocks.java
@@ -1,0 +1,20 @@
+package com.mygdx.mathdash.sprites.tileObjects;
+
+import com.badlogic.gdx.Gdx;
+import com.badlogic.gdx.maps.tiled.TiledMap;
+import com.badlogic.gdx.math.Rectangle;
+import com.badlogic.gdx.physics.box2d.World;
+import com.mygdx.mathdash.MathDash;
+
+public class FinishBlocks extends InteractiveTileObject{
+    public FinishBlocks(World world, TiledMap map, Rectangle bounds) {
+        super(world, map, bounds);
+        fixture.setUserData(this);
+        setCateforyFilter(MathDash.FINISH_BLOCK_BIT);
+    }
+
+    @Override
+    public void contacting() {
+        Gdx.app.log("Финиш", "Конец");
+    }
+}
