+++ core/src/com/mygdx/mathdash/sprites/tileObjects/InteractiveTileObject.java
@@ -1,0 +1,49 @@
+package com.mygdx.mathdash.sprites.tileObjects;
+
+import com.badlogic.gdx.maps.tiled.TiledMap;
+import com.badlogic.gdx.math.Rectangle;
+import com.badlogic.gdx.physics.box2d.Body;
+import com.badlogic.gdx.physics.box2d.BodyDef;
+import com.badlogic.gdx.physics.box2d.Filter;
+import com.badlogic.gdx.physics.box2d.Fixture;
+import com.badlogic.gdx.physics.box2d.FixtureDef;
+import com.badlogic.gdx.physics.box2d.PolygonShape;
+import com.badlogic.gdx.physics.box2d.World;
+import com.mygdx.mathdash.MathDash;
+
+public abstract class InteractiveTileObject {
+    protected World world;
+    protected TiledMap map;
+    protected Rectangle bounds;
+    protected Body body;
+
+    protected Fixture fixture;
+
+    public InteractiveTileObject(World world, TiledMap map, Rectangle bounds) {
+        this.world = world;
+        this.map = map;
+        this.bounds = bounds;
+
+        BodyDef bdef = new BodyDef();
+        FixtureDef fdef = new FixtureDef();
+        PolygonShape shape = new PolygonShape();
+
+        bdef.type = BodyDef.BodyType.StaticBody;
+        bdef.position.set((bounds.getX() + bounds.getWidth() / 2) / MathDash.PPM, (bounds.getY() + bounds.getHeight() / 2) / MathDash.PPM);
+
+        body = world.createBody(bdef);
+
+        shape.setAsBox(bounds.getWidth() / 2 / MathDash.PPM, bounds.getHeight() / 2 / MathDash.PPM);
+        fdef.shape = shape;
+        body.createFixture(fdef);
+        fixture = body.createFixture(fdef);
+    }
+
+    public abstract void contacting();
+    public void setCateforyFilter(short filterBit) {
+        Filter filter = new Filter();
+        filter.categoryBits = filterBit;
+        fixture.setFilterData(filter);
+    }
+
+}
