--- desktop/src/com/mygdx/game/DesktopLauncher.java
@@ -1,14 +1,0 @@
-package com.mygdx.game;
-
-import com.badlogic.gdx.backends.lwjgl3.Lwjgl3Application;
-import com.badlogic.gdx.backends.lwjgl3.Lwjgl3ApplicationConfiguration;
-
-// Please note that on macOS your application needs to be started with the -XstartOnFirstThread JVM argument
-public class DesktopLauncher {
-	public static void main (String[] arg) {
-		Lwjgl3ApplicationConfiguration config = new Lwjgl3ApplicationConfiguration();
-		config.setForegroundFPS(60);
-		config.setTitle("Geometry Dash 2.0");
-		new Lwjgl3Application(new MathDash(), config);
-	}
-}
