+++ ios/build.gradle
@@ -4,7 +4,7 @@ sourceCompatibility = '1.7'
 [compileJava, compileTestJava]*.options*.encoding = 'UTF-8'
 
 ext {
-	mainClassName = "com.mygdx.game.IOSLauncher"
+	mainClassName = "com.mygdx.mathdash.IOSLauncher"
 }
 
 launchIPhoneSimulator.dependsOn build
