+++ ios/robovm.properties
@@ -1,5 +1,5 @@
 app.version=1.0
-app.id=com.mygdx.game
+app.id=com.mygdx.mathdash
 app.mainclass=com.mygdx.game.IOSLauncher
 app.executable=IOSLauncher
 app.build=1
